training/requirements.txt
asn1crypto==0.24.0
astroid==2.1.0
autopep8==1.4.3
bcrypt==3.1.6
boto3==1.9.86
botocore==1.12.86
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.5
docutils==0.14
idna==2.8
isort==4.3.4
jmespath==0.9.3
lazy-object-proxy==1.3.1
mccabe==0.6.1
ncloud-apikey==1.0.5
ncloud-autoscaling==1.0.7
ncloud-cdn==1.0.7
ncloud-clouddb==1.0.7
ncloud-loadbalancer==1.0.7
ncloud-monitoring==1.0.7
ncloud-sdk==1.0.10
ncloud-server==1.0.7
netifaces==0.10.9
paramiko==2.4.2
pyasn1==0.4.5
pycodestyle==2.5.0
pycparser==2.19
pylint==2.2.2
PyNaCl==1.3.0
python-dateutil==2.7.5
pyzmq==17.1.2
requests==2.21.0
s3transfer==0.1.13
six==1.12.0
urllib3==1.24.1
wrapt==1.11.1
